+++ fixtures/languages/test.jsx
@@ -1,10 +1,7 @@
-// JavaScript React file
-// JSX component
 function Component() {
-  // Inside component comment
   return (
     <div>
-      {/* JSX comment */}
+      {}
       Hello World
     </div>
   );