+++ fixtures/languages/test.php
@@ -1,0 +1,5 @@
+<?php
+
+// This comment should be removed
+$s = "// not a comment";
+echo $s;