+++ fixtures/languages/test.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# This comment should be removed
+puts "Hello # not a comment"